requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
filelock==3.12.2
Pillow==10.0.0
platformdirs==3.10.0
pymongo==4.5.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.2
